Refact keyboard.c mouse.c
