+++ zmosII/harib00j/Makefile
@@ -35,9 +35,9 @@ asmhead.bin : asmhead.nas Makefile
 
 
 
-bootpack.bim : bootpack.obj graphic.obj dsctbl.obj int.obj fifo.obj naskfunc.obj hankaku.obj Makefile
+bootpack.bim : bootpack.obj graphic.obj dsctbl.obj int.obj fifo.obj naskfunc.obj mouse.obj keyboard.obj hankaku.obj Makefile
 	$(OBJ2BIM) @$(RULEFILE) out:bootpack.bim stack:3136k map:bootpack.map \
-		bootpack.obj graphic.obj dsctbl.obj int.obj fifo.obj naskfunc.obj hankaku.obj
+		bootpack.obj graphic.obj dsctbl.obj int.obj fifo.obj naskfunc.obj mouse.obj keyboard.obj hankaku.obj
 # 3MB+64KB=3136KB
 
 bootpack.hrb : bootpack.bim Makefile
