+++ zmosII/harib00j/bootpack.c
@@ -1,8 +1,6 @@
 #include <stdio.h>
 #include "bootpack.h"
 
-extern struct FIFO8 keyfifo;
-extern struct FIFO8 mousefifo;
 
 
 void HariMain(void)
@@ -15,17 +13,18 @@ void HariMain(void)
 	
 	init_gdtidt();
 	init_pic();
+	io_sti();
+		
+	fifo8_init(&keyfifo,keybuff,36);
+	fifo8_init(&mousefifo,mousebuff,36);
 	
 	/*由于 init_pic的时候 禁用了所有IRQ，这里需要手动开放需要的IRQ */
 	io_out8(PIC0_IMR, 0xf9); /* PIC0(主PIC) 开放IRQ-1(键盘) IRQ-2(链接 从PIC) (11111001) */
 	io_out8(PIC1_IMR, 0xef); /* PIC1(从PIC) 开放IRQ-12(鼠标) (11101111) */
-	
 	init_keyboard();
 	enable_mouse(&mdec);
-	io_sti();
 	init_palette();
 	init_screen(binfo);
-	
 	
 	/* init mouse start */
 	int mx,my;
@@ -36,17 +35,8 @@ void HariMain(void)
 	putblock8_8(binfo->vram,binfo->scrnx,16,16,mx,my,cursor,16);
 	/* init mouse end */
 	
-	
-	fifo8_init(&keyfifo,keybuff,36);
-	fifo8_init(&mousefifo,mousebuff,36);
-	
-	
-	
-	
-	
 	unsigned data;
 	int keyBufDataIndex;
-	
 	unsigned char mouse_data_buf[3], mouse_phase;
 	mouse_phase = 0;
 	while(1){
@@ -92,79 +82,5 @@ void HariMain(void)
 			}
 		}
 	}
-}
-
-
-
-
-
-void wait_KBC_sendready(void)
-{
-	for(;;){
-		/*io_in8(PORT)KEYSTA) will return 0x10 If keyboard is not reay*/
-		if( (io_in8(PORT_KEYSTA) & KEYSTA_SEND_NOTREADY) == 0){
-			break;
-		}
-	}
-	return;
-}
-
-
-void init_keyboard(void)
-{
-	wait_KBC_sendready();
-	io_out8(PORT_KEYCMD,KEYCMD_WRITE_MODE);
-	wait_KBC_sendready();
-	io_out8(PORT_KEYDAT,KBC_MODE);
-	return;
-}
-
-void enable_mouse(struct MOUSE_DESC *mdec)
-{
-	wait_KBC_sendready();
-	io_out8(PORT_KEYCMD,KEYCMD_SENDTO_MOUSE);
-	wait_KBC_sendready();
-	io_out8(PORT_KEYDAT,MOUSECMD_ENABLE);
-	mdec->phase = 0;
-	return;
-}
-
-
-int mouse_decode(struct MOUSE_DESC *mdec, unsigned char data)
-{
-		if(mdec->phase == 0){
-			if(data == 0xfa){
-				mdec->phase = 1;
-				return 0;
-			}
-		}else if(mdec->phase == 1){
-			if( (data & 0xc8) == 0x08){
-				mdec->buf[0] = data;
-				mdec->phase = 2;
-			}
-			return 0;
-	    }else if(mdec->phase == 2){
-	    	mdec->buf[1] = data;
-	    	mdec->phase = 3;
-			return 0;
-	    }else if(mdec->phase == 3){
-	    	mdec->buf[2] = data;
-	    	mdec->phase = 1;
-			/*开始解析鼠标的信息*/
-			mdec->btn = mdec->buf[0] & 0x07; //0x07 = 0000_0111B,鼠标状态被存放在buf[0]的低三位。
-			mdec->x = mdec->buf[1];
-			mdec->y = mdec->buf[2];
-			
-			if( (mdec->buf[0] & 0x10) != 0){
-				mdec->x |= 0xffffff00;
-			}
-			if( (mdec->buf[0] & 0x20) != 0){
-				mdec->y |= 0xffffff00;
-			}
-			
-			mdec->y = -mdec->y;
-			return 1;
-		}
-	return -1;
 }
