+++ zmosII/harib00j/int.c
@@ -2,8 +2,8 @@
 
 
 
-struct FIFO8 keyfifo;
-struct FIFO8 mousefifo;
+
+
 /*
  *初始化PIC，固定配置
 */
@@ -29,24 +29,3 @@ void init_pic(void)
 	return;
 }
 
-//Keyboard
-void inthandler21(int *esp)
-{
-	unsigned char data;
-	io_out8(PIC0_OCW2,0x60 + 1); //通知 主PIC IRQ-1 中断处理完毕。
-	data = io_in8(PORT_KEYDAT);
-	fifo8_put(&keyfifo,data);
-	return;
-}
-
-void inthandler2c(int *esp)
-{
-	unsigned char data;
-	io_out8(PIC1_OCW2,0x60 + 4); //通知 从PIC IRQ-12 中断处理完毕。
-	io_out8(PIC0_OCW2,0x60 + 2); //通知 主PIC IRQ-2 中断处理完毕。
-	
-	data = io_in8(PORT_KEYDAT);
-	fifo8_put(&mousefifo,data);
-	
-	return;
-}
